refactor: generalize use polling reload to be reused
