+++ identity/client/src/pages/cluster-variables/List.tsx
@@ -21,7 +21,9 @@ import { AddModal } from "./modals/add-modal";
 import DeleteModal from "./modals/DeleteModal";
 import DetailsModal from "./modals/DetailsModal";
 import EditModal from "./modals/EditModal";
-import { usePollingReload } from "./usePollingReload";
+import { usePollingReload } from "src/utility/hooks/usePollingReload";
+import { type QueryClusterVariablesResponseBody } from "@camunda/camunda-api-zod-schemas/8.9";
+import { useCallback } from "react";
 
 export default function List() {
   const { t } = useTranslate("clusterVariables");
@@ -34,14 +36,39 @@ export default function List() {
     ...paginationProps
   } = usePaginatedApi(searchClusterVariables);
 
+  const compareClusterVariables = useCallback(
+    (current: QueryClusterVariablesResponseBody) => {
+      const previous = clusterVariables?.items ?? [];
+      if (previous.length !== current.items.length) {
+        return true;
+      }
+
+      return previous.some((prev) => {
+        const match = current.items.find(
+          (curr) =>
+            curr.name === prev.name &&
+            curr.scope === prev.scope &&
+            curr.tenantId === prev.tenantId,
+        );
+        if (!match) {
+          return true;
+        }
+        return JSON.stringify(prev.value) !== JSON.stringify(match.value);
+      });
+    },
+    [clusterVariables],
+  );
+
   const { startPolling, pollingStatus, resetPollingStatus, isPolling } =
-    usePollingReload(reload);
+    usePollingReload<QueryClusterVariablesResponseBody>(
+      reload,
+      compareClusterVariables,
+    );
 
   const isInitialLoad = loading && !isPolling;
 
   const reloadWithPolling = () => {
-    const snapshot = clusterVariables?.items ?? [];
-    startPolling(snapshot);
+    startPolling();
   };
 
   const [addClusterVariable, addClusterVariableModal] = useModal(
